vendors: README
The folders in vendors come from outside applications and they are too
big (as in several dozens MB) to be added to our repository. It's much
simpler to list them and to go download them (either from their original
websites or we could put a zip somewhere in bdovore.com.